Finished items 1-4.
I'm not sure if my media queries are working correctly - I looked them over several times and the syntax seems right, but I'm only able to see the background color change to pink or green depending on the size of my browser window. The third color I expected to see was yellow.